feat: replace setting button with reusable BottomRightButton
