+++ lib/constants/dimensions.dart
@@ -1,2 +1,6 @@
+import 'dart:ui';
+
 const double iconSize = 20.0;
 const double iconButtonSize = 44.0;
+const Size bottomRightButtonSize = Size(48, 48);
+const Offset bottomRightButtonOffset = Offset(32, 40);
